debug steps, check if symlink exists

+++ R/scripts/check.sh.tpl
@@ -92,7 +92,9 @@ mkdir -p "${TMP_HOME}"
 export HOME="${TMP_HOME}"
 
 if [[ "{pkg_src_archive}" != "{pkg_name}.tar.gz" ]]; then
-  ln -s "{pkg_src_archive}" "{pkg_name}.tar.gz"
+  pwd
+  ls -la
+  test -L "${pkg_name.tar.gz}" || ln -s "{pkg_src_archive}" "{pkg_name}.tar.gz"
 fi
 
 {R} CMD check {check_args} "{pkg_name}.tar.gz"
